Move linfa-nn main tests into integration tests
